# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Theory of Change API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /theories:
    get:
      operationId: listTheories
      tags: [theories]
      summary: List all theories of change
      responses:
        "200":
          description: List of theories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Theory"
    post:
      operationId: createTheory
      tags: [theories]
      summary: Create a new theory of change
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTheory"
      responses:
        "201":
          description: Created theory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Theory"
  /theories/{id}:
    get:
      operationId: getTheory
      tags: [theories]
      summary: Get a theory of change
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Theory details with components
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TheoryDetail"
        "404":
          description: Not found
    put:
      operationId: updateTheory
      tags: [theories]
      summary: Update a theory of change
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTheory"
      responses:
        "200":
          description: Updated theory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Theory"
    delete:
      operationId: deleteTheory
      tags: [theories]
      summary: Delete a theory of change
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /theories/{theoryId}/components:
    get:
      operationId: listComponents
      tags: [components]
      summary: List components for a theory
      parameters:
        - name: theoryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of components
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Component"
    post:
      operationId: createComponent
      tags: [components]
      summary: Create a component in a theory
      parameters:
        - name: theoryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateComponent"
      responses:
        "201":
          description: Created component
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Component"
  /theories/{theoryId}/components/{id}:
    put:
      operationId: updateComponent
      tags: [components]
      summary: Update a component
      parameters:
        - name: theoryId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateComponent"
      responses:
        "200":
          description: Updated component
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Component"
    delete:
      operationId: deleteComponent
      tags: [components]
      summary: Delete a component
      parameters:
        - name: theoryId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /theories/{theoryId}/components/{componentId}/indicators:
    get:
      operationId: listComponentIndicators
      tags: [components]
      summary: List indicators for a component
      parameters:
        - name: theoryId
          in: path
          required: true
          schema:
            type: integer
        - name: componentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of indicators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ComponentIndicator"
    post:
      operationId: createComponentIndicator
      tags: [components]
      summary: Create an indicator for a component
      parameters:
        - name: theoryId
          in: path
          required: true
          schema:
            type: integer
        - name: componentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateComponentIndicator"
      responses:
        "201":
          description: Created indicator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComponentIndicator"
  /theories/{theoryId}/components/{componentId}/indicators/{id}:
    put:
      operationId: updateComponentIndicator
      tags: [components]
      summary: Update an indicator
      parameters:
        - name: theoryId
          in: path
          required: true
          schema:
            type: integer
        - name: componentId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateComponentIndicator"
      responses:
        "200":
          description: Updated indicator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComponentIndicator"
    delete:
      operationId: deleteComponentIndicator
      tags: [components]
      summary: Delete an indicator
      parameters:
        - name: theoryId
          in: path
          required: true
          schema:
            type: integer
        - name: componentId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /theories/{theoryId}/connections:
    get:
      operationId: listConnections
      tags: [connections]
      summary: List connections between components
      parameters:
        - name: theoryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of connections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Connection"
    post:
      operationId: createConnection
      tags: [connections]
      summary: Create a connection between components
      parameters:
        - name: theoryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConnection"
      responses:
        "201":
          description: Created connection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
  /theories/{theoryId}/connections/{id}:
    patch:
      operationId: updateConnection
      tags: [connections]
      summary: Update a connection (e.g. anchor points)
      parameters:
        - name: theoryId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateConnection"
      responses:
        "200":
          description: Updated connection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
    delete:
      operationId: deleteConnection
      tags: [connections]
      summary: Delete a connection
      parameters:
        - name: theoryId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /theories/{theoryId}/risk-analyses:
    get:
      operationId: listTheoryRiskAnalyses
      tags: [theories]
      summary: List risk analyses for a theory
      parameters:
        - name: theoryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of risk analyses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TheoryRiskAnalysis"
    post:
      operationId: createTheoryRiskAnalysis
      tags: [theories]
      summary: Create a risk analysis entry
      parameters:
        - name: theoryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTheoryRiskAnalysis"
      responses:
        "201":
          description: Created risk analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TheoryRiskAnalysis"
  /theories/{theoryId}/risk-analyses/{id}:
    patch:
      operationId: updateTheoryRiskAnalysis
      tags: [theories]
      summary: Update a risk analysis entry
      parameters:
        - name: theoryId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTheoryRiskAnalysis"
      responses:
        "200":
          description: Updated risk analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TheoryRiskAnalysis"
    delete:
      operationId: deleteTheoryRiskAnalysis
      tags: [theories]
      summary: Delete a risk analysis entry
      parameters:
        - name: theoryId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /theories/{theoryId}/notes-updates:
    get:
      operationId: listTheoryNotesUpdates
      tags: [theories]
      summary: List notes and updates for a theory
      parameters:
        - name: theoryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of notes and updates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TheoryNoteUpdate"
    post:
      operationId: createTheoryNoteUpdate
      tags: [theories]
      summary: Create a note or update
      parameters:
        - name: theoryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTheoryNoteUpdate"
      responses:
        "201":
          description: Created note/update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TheoryNoteUpdate"
  /theories/{theoryId}/notes-updates/{id}:
    patch:
      operationId: updateTheoryNoteUpdate
      tags: [theories]
      summary: Update a note or update entry
      parameters:
        - name: theoryId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTheoryNoteUpdate"
      responses:
        "200":
          description: Updated note/update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TheoryNoteUpdate"
    delete:
      operationId: deleteTheoryNoteUpdate
      tags: [theories]
      summary: Delete a note or update entry
      parameters:
        - name: theoryId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /theories/{theoryId}/change-log:
    get:
      operationId: listTheoryChangeLog
      tags: [theories]
      summary: List change log entries for a theory
      parameters:
        - name: theoryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of change log entries, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChangeLogEntry"
...
components:
  schemas:
...
    Theory:
      type: object
      properties:
        id:
          type: integer
        portfolioId:
          type: integer
          nullable: true
        title:
          type: string
        description:
          type: string
        market:
          type: string
        interventionCode:
          type: string
        interventionTitle:
          type: string
        manager:
          type: string
        mrmResponsible:
          type: string
        targetBeneficiary:
          type: string
        privateSectorPartners:
          type: string
        publicSectorPartners:
          type: string
        serviceProviders:
          type: string
        strategy:
          type: string
        interventionStory:
          type: string
        womenEconomicEmpowerment:
          type: string
        climateSmart:
          type: string
        displacement:
          type: string
        contributionOfOtherProjects:
          type: string
        businessModelImagePath:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - createdAt
        - updatedAt
...
    Connection:
      type: object
      properties:
        id:
          type: integer
        theoryId:
          type: integer
        fromComponentId:
          type: integer
        toComponentId:
          type: integer
        label:
          type: string
        startAnchor:
          type: string
          nullable: true
        endAnchor:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - theoryId
        - fromComponentId
        - toComponentId
        - label
        - createdAt
    CreateConnection:
      type: object
      properties:
        fromComponentId:
          type: integer
        toComponentId:
          type: integer
        label:
          type: string
      required:
        - fromComponentId
        - toComponentId
        - label
...